Update with beta4 features
