+++ package.json
@@ -15,13 +15,13 @@
     "test.watch": "jest --watch --no-cache"
   },
   "dependencies": {
-    "@bearer/core": "0.71.4",
-    "@bearer/intents": "0.71.0",
-    "@bearer/ui": "0.71.4",
+    "@bearer/core": "next",
+    "@bearer/intents": "next",
+    "@bearer/ui": "next",
     "axios": "^0.18.0"
   },
   "devDependencies": {
-    "@bearer/cli": "~0.71.4",
+    "@bearer/cli": "next",
     "@types/jest": "^23.3.5",
     "@types/node": "^10.11.0",
     "jest": "latest"
